Fix rendering of polygons without stroke.
Looking at the call signature of
appendFlatCoordinates(flatCoordinates, offset, end, stride, closed, skipFirst)
and the comments in PolygonReplay.js#44 it seem that the !stroke argument should rather be the "closed" argument to appendFlatCoordinates() than the "skipFirst" argument. "skipFirst" "hard-skips" the first vertex of the polygon. This turns quadrilaterals into triangles and so on.